backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
yfinance
pandas
numpy
scikit-learn
ta
apscheduler
python-dotenv
setuptools
